packages/opentelemetry-instrumentation-azure-search/pyproject.toml
[tool.coverage.run]
branch = true
source = ["opentelemetry/instrumentation/azure-search"]

[tool.coverage.report]
exclude_lines = ["if TYPE_CHECKING:"]
show_missing = true

[tool.poetry]
name = "opentelemetry-instrumentation-azure-search"
version = "0.1.0"
description = "OpenTelemetry instrumentation for Azure AI Search"
authors = ["Ali Al-Bajati <ali.oa.albajati@gmail.com>"]
repository = "https://github.com/traceloop/openllmetry/tree/main/packages/opentelemetry-instrumentation-azure-search"
license = "Apache-2.0"
readme = "README.md"

[[tool.poetry.packages]]
include = "opentelemetry/instrumentation/azure-search"

[tool.poetry.dependencies]
python = ">=3.9,<4"
opentelemetry-api = "^1.28.0"
opentelemetry-instrumentation = ">=0.50b0"
opentelemetry-semantic-conventions = ">=0.50b0"
opentelemetry-semantic-conventions-ai = "0.4.2"
tiktoken = ">=0.6.0, <1"
azure-search-documents = "^11.5.3"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.2.0"
flake8 = "7.0.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.2.2"
pytest-sugar = "1.0.0"
vcrpy = "^6.0.1"
pytest-recording = "^0.13.1"
azure-search-documents = "^11.4.0"
opentelemetry-sdk = "^1.27.0"
pytest-asyncio = "^0.23.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.extras]
instruments = ["azure-search-documents"]

[tool.poetry.plugins."opentelemetry_instrumentor"]
azure-search = "opentelemetry.instrumentation.azure-search:AzureSearchInstrumentor"